feat: update function parser
